change variable names for more clarity
Generalize the name convention changing 'self.eos' to 'self.classes' and all local 'eos' to 'clas'. CoReDataset still treats each EOS as a class (category), which is indicated in the class docstring, but this is only at the instance initialization. The rest of the class is agnostic to this choice, so it can be changed / generalized later if needed.