hide PSD/ASD interpolators and increase their numerical representation's resolution
The interpolators will be now stored at _psd and _asd attributes.
Also, in order to prevent loss of accuracy as much as possible when pickling instances,
the frequency resolution of the arrays has been increased to 16 points per frequency bin. Furthermore, when unpickling an instance, the array representation will be kept the same, so no further data degradation will be produced due to precision errors.