use the ID list from 'self.labels' instead of 'self.metadata.index'
The Base class was initially conceived to require the metadata to be stored in a Pandas DataFrame in the 'metadata' attribute. However, there could be cases where there is no need for metadata (such as the upcoming class UnlabeledWaves). This change allows the possibility of creating a class inheriting Base, but not including the 'metadata' attribute.

WARNING: At the time of this commit, it's not been tested the small possibility of this introducing unexpected behavior in CoReWaves and InjectedCoReWaves. In principle, the IDs inside the dictionary 'self.labels' should follow the same order as in 'self.metadata.index'. However, 'self.labels' is built from the first 2 layers in 'self.strains', and hence there is no explicit enforcement of these following the same order as in 'self.metadata.index'. If the order is not conserved, I am not sure whether this could affect the train/test splitting w.r.t. saved datasets generated using previous versions.