fix inconsistence in `flength`
There is still an inconsistence issue regarding the highpass filter applied to the data and to generate the noise. Initially I allowed different values to be set for each of these processes. I am not sure currently if this has any use, but the main issue is that at some point I began to confuse their purpose. I've fixed some calls, but I still need to do a proper revision and potential rewrite of these parameters.