hide the dummy class name from the user in Unlabeled-like classes
For now I think it's more practical to implement the unlabeled classes as a specific case of the base classes using dummy class and labels. I've gathered the methods that needed to be rewritten from the Base and BaseInjected classes into a mix-in class which must be inherited along Base (or BaseInjected) with higher priority in the MRO.

I fear that in the long run this could over-complicate escalation. Another option could be to first write a base class for the more general case of Unlabeled signals, and then implement by inheritance the labeled and class cases. This approach, however, implies rewriting a good chunk of code. Therefore I'll avoid it so long as the current approach proves to be reasonably easy to escalate.