don't make the dummy class invisible
Public methods like `items()` and `keys()` are used by internal methods as well, and expect the 1st "layer" to be a class, even if it's the dummy label. In order to keep the code clean and simple, I will not omit this layer, and just add the option for some retrieval operations to omit the 1st parameter, as this can easily be accounted for in the case of unlabeled datasets (the dummy class will always be labeled 'unique').