patch temporarily the initial estimation of the merger (+info)
CoRe strains don't usually come with the simulation's time centered around the merger. When initialising CoReWaves, even though I was calling the find_time_origin method, the initial value of `merger_pos` was wrong. It happened to be fixed after calling `project`, because it made use of `tat.find_merger`, which actually looks at the strain's amplitude to roughly estimate the position of the merger, and with it update the time arrays.

TODO: Make this process more robust by either adding an argument to the __init__ method, or (and) checking the agreement between the time origin and the estimated merger position.