optimise Base.whiten method and fix some minor issues (+info)
Now when shrink is given, the method tries to do a first shrinking (after estimating the individual ASD if necessary) to avoid whitening more than needed. This takes into account the filter settle-in window. After the whitening, the method removes the settle-in margins left for whitening.

Minimal details were updated in BaseInjected.whiten, but it still requires a full check.