fix shrinking of original copies for the shrink paths (+info)
When 1 or more whitened strains where shrunk before whitening, these changes were not propagated to the original copies. At the same time, if there were shrink_later strains, these were indeed propagated, leading to an inconsistent state. With the new option implemented in the shrinking method, we now perform all shrinking operations of the whiten method on the current strains only, and leave the propagation to the original strains to the side-effect hook.